chore: port node transform and outline visual tests
